server/prospekt_pipeline/requirements.txt
beautifulsoup4>=4.12.0
pypdfium2>=4.18.0
pdf2image>=1.16.0
pillow>=10.0.0
pytesseract>=0.3.10
paddleocr>=2.7.0
thefuzz>=0.19.0
rapidfuzz>=3.0.0
openai>=1.0.0
python-dotenv>=1.0.0
pdfminer.six>=20221105
PyPDF2>=3.0.0
tqdm>=4.66.0
pytest>=7.4.0

